Fixes stdin redirection issues within the command-line module.
The callback passed into `command-line.run` would return the argument list:
`stderr, stdout`, however, these were not populated in the command-line module
as `stdio` was set to 'inherit' in the `spawn` function call.

This meant that the `error` argument of the callback would always be a falsey
empty string.  The result of this was that the sass gem was always downloaded
and installed each time `origami install`. The `stdout` stream from
`sass --version` was being sent directly to the real 'stdout'.

This alters the API providing a single point for spawning processes which adds
an optional `option` argument for controlling stderr and stdout.