Support loading of local partials in demos
Mustache files in the same directory as the demo template are
automatically loaded in and set as partials. This works for files in
subdirectories and allows for reuse of HTML in the demos.

This does not support full path resolution, and will not find mustache
files in directories further up the tree.