Replace babel presets with preset-env
This is after running tests to determine that the output is identical for
es2015,2016,2016 and es3 code (output string is deepEquals).

es3 presets are _not_ included in preset-env, but without configuration it is identical
to the es-preset-20xx modules.

It would be nice to have some form of snapshot testing to verify this.