Update the default support details (#570)
Currently `obt init` generates an `origami.json` file that specifies a
support status of "active" and support contact details which point to
the Origami team.

This means that the responsibility for marking a new component as
experimental or maintained by another team is on the person creating a
component. This makes work for us, as we have to remind a non-Origami
developer to update these fields before publishing, and it's possible
(even likely) that an unsupported component will appear in the registry
as "actively maintained by the Origami team".

This PR updates the template to provide sensible details for users who
aren't as familiar with Origami. Members of the Origami team who create
new components are more likely to check and update this information as
required without prompting.