Test dart and node sass compilation. (#593)
* Test dart and node sass compilation.

As Dart Sass is the primary implementation of Sass, we look to
use Dart Sass in a future major of obt or [oat](https://github.com/financial-times/origami-authoring-tools).

This commit adds Dart Sass compilation tests so we can be sure
our components compile with both Node and Dart Sass.

* Pass dart-sass library to the build function.

- Allow "outputStyle" to be customised -- Dart Sass does not support "nested".
- Remove custom "@warn" function from compilation test. This is not
compatible with Dart Sass, which can add custom functions but not custom
annotations. This means warnings are logged as there is no '--quiet' option.