migrate from serve to latest serve-handler (#687)
serve no longer has a API, it is only a terminal command now. I've moved us to the API package which is called serve-handler.

I've kept the same look and feel of serve but not added the copy link to clipboard feature. I can add that if we want it.