Remove remote demo data feature.
- Our components do not use remote demo data.
- o-header could use this feature to show up-to-date links from
the Navigation Service, but if users copy/paste their navigation
will become out of date. Instead we should use dummy data
(https://github.com/Financial-Times/o-header/issues/160)
- The Origami specification says "if [the demo field] is a string
it must be a path to a JSON file containing the data, relative to
the root of the repo." So remote data is not allowed by the spec.

(Let's not worry about the effort I went to in a previous commit,
to replace `request-promise-native` with `node-fetch`.)