Update tests to check tab indent linting.
The spec says component indent type (tabs or spaces) is not
standardised: developers must respect whatever indent type is already
in use when editing existing components.
https://origami.ft.com/spec/v1/sass/#syntax-convention

In practise that’s quite hard and we end up using both within a
project. So we want to lint that tabs are used, unless spaces
are configured.