Add demo command tests and fix a related bug.
The polyfill url in demos includes the `default` set, which we
had intended to remove to ensure we catch errors where a
JavaScript feature is not supported by the browser. Only polyfills
which have been explicitly specified in `origami.json` should be
included in demos.

However we only removed the `default` set for Karma tests by mistake:
https://github.com/Financial-Times/origami-build-tools/pull/623