Update stylelint/esline ignore paths.
Use the component's `.gitignore` instead of parsing `.gitignore`
paths within origami build tools. This means `.eslintignore` is
not used and there is no fallback if `.gitignore` is not specified.

The latter is not a problem, the former maybe? The only component
which currently uses `.eslintignore` is `o-tracking`, and those
are old/`.gitignore` duplicates:
https://github.com/Financial-Times/o-tracking/blob/04862efac02f7fc952c505b9335a92b5166c0b34/.eslintignore#L2