Use a `.gitignore` eslint ignore path to exclude `bower_components`
Fixes component JS linting which currenly returns many errors
outside the component source.

Using `.gitignore` means `.eslintignore` is not used. There is
no fallback if `.gitignore` is not specified.

The latter is not a problem. The former maybe? The only component
which currently uses `.eslintignore` is `o-tracking`, and those
are old/`.gitignore` duplicates:
https://github.com/Financial-Times/o-tracking/blob/04862efac02f7fc952c505b9335a92b5166c0b34/.eslintignore#L2

The alternative is to restore obt parsing of `.gitignore` to pass
as additional ignore patterns:
https://github.com/Financial-Times/origami-build-tools/pull/926/commits/96fea8e87fcc7fc4eb50a31efe158e5bb2acceb2#diff-a99c9b5d896f4bcdda8aaba5b441d6d7771390c07e2b54631eb403e1e77eddecL19