Fix component Sass linting
Stylelint plugins require `stylelint` which means if we’re using
component lint config and obt doesn’t depend on the same plugins
the plugin isn't resolved and an error is thrown

To avoid this we can run the stylelint cli with `npx`. This
feels not great but has some advantages. If a component depends
on and configures a newer version of Stylelint obt won't error.

In the future perhaps the component spec should should define
`lint` as an expected `package.json` script which obt could defer
to.