Make git pushes in create-new-version workflow trigger other workflows
Currently we are using the read-only github token. When using this token and pushing commits/tags to github, github will not trigger any workflows for the pushed commit/tag.

We need the pushed tag to trigger out npm-publishing workflows, which means we need to use a custom github token instead of the default token provided by github actions.