Use Dart Sass in tests
Karma is currently using node-sass which may cause errors when
running `npm test` if the component under test relied on dart
sass features (e.g. relies on `_index.scss` to resolve imports)

Currently relies on an embedded, modified version of
`karma-scss-preprocessor`. There is an upstream PR which we
may be able to switch to if it's accepted:
https://github.com/amercier/karma-scss-preprocessor/pull/73