ensure npm 7 is used when running obt install
obt install is set up to use a locally installed version of npm if one exists. A local version of npm die not exist within the project and so obt install would fallback to using whatever `npm` would resolve to from the users PATH. This can cause issues for users if their global npm command is an old version of npm for example (This happened to me today).

By installing npm as a dependency to obt, we snure that obt install will always use the version of npm that we want, which is npm version 7.