Remove last few remaining references to bower
This removes bower_components from the ignored filters for `obt dev`, `obt verify`, and the projects .gitignore.

This removes references to the no longer existing `ingoreBower` configuration option which was left behind in the pa11y task and test sass task.

This also removes the last remaining test which involved a bower dependency.