fix bug in karma config due to misuing the karma api
Newer versions of Karma throw an error if a file is supplied as a string instead of a file object.

Here we have updated our config to use the documented file object to ensure the newer versions of karma work with obt.

This is currently an issue everyone will hit if they are installing origami-build-tools for the first time because newer versions of karma will get installed (because we do not include an npm-shrinkwrap).