Add validation for package.json.browser field
The component v2 specification states that if a main.js file exists, then the `browser` field must be set to `"main.js"`. This commit updates the verify task to ensure that the component being tested adheres to these rules.

I've quoted below the parts of the specification this is testing.

>  - It **must** include a <a href="https://docs.npmjs.com/cli/v7/configuring-npm/package-json#browser">`browser`</a> property set to the component's main JavaScript file (`main.js`) **_if_** it exists.

Those came from the spec v2 pull-request -- https://github.com/Financial-Times/origami-website/pull/273/files#diff-5ebc2f3b798414ec905f3fdceb910557607b8d779a56bf1ce09ad48e32fd085aR88

I've also added in a check for the opposite situation, if no `main.js` file exists, then the `browser` field should not exist. This is not currently written in the specification though. If we don't want this functionality, I will remove it. If we do want this functionality, I will update the specification to reflect this.