Skip pa11y tests when the `--ignore-bower` flag is set.
`origami-ci-tools` runs `obt test --ignore-bower`, to test the
npm package of a component which supports Bower and npm is correct:
https://github.com/Financial-Times/origami-ci-tools/blob/master/commands/branch.js#L52

This used to not run pa11y tests as the pa11y demo is not built by
`origami-ci-tools`, and Origami Build Tools used to skip pa11y tests
if a built pa11y demo was not found. However origami-build-tools
was updated to check for pa11y source code and build the pa11y
demo if it exists:
https://github.com/Financial-Times/origami-build-tools/pull/933/files#diff-6c44fc3cf122c89eee3ab192bd97d1a853d25adb56fa750332e79aac7ad230d4R90

`origami-component-converter`, used to generate the npm bundle, does
not support demo source code so the pa11y build fails, and thus
`origami-ci-tools` fails, and thus we can't make minor/patch release
to old versions of components which still support Bower.

By skipping pa11y tests if `--ignore-bower` is set we avoid this
error.