Use fs.access instead of fs.open to avoid having to open and close file-descriptors.
NodeJS is deprecating the automatic closing of fileHandlers during garbage collection.

This commit updates the codebase to explicitly close the fileHandlers instead of relying on the deprecated behaviour.

Below is the message that is displayed when running obt prior to this commit:

"DeprecationWarning: Closing a FileHandle object on garbage collection is deprecated. Please close FileHandle objects explicitly using FileHandle.prototype.close(). In the future, an error will be thrown if a file descriptor is closed during garbage collection."